refactor: switch to realtimekit core
